+++ src/main/java/com/example/demo/repository/LikeRepository.java
@@ -10,12 +10,12 @@ import com.example.demo.model.Like;
 @Repository
 public interface LikeRepository extends JpaRepository<Like, Long> {
 	
-	public List<Like> findByPostId(Long postId);
-	public int countByPostId(Long postId);
+	public List<Like> findByPostId(int postId);
+	public int countByPostId(int postId);
 	
 	//既にいいねしているか検索
-	boolean existsByUsernameAndPostId(String username, Long postId);
+	boolean existsByUsernameAndPostId(String username, int postId);
 	
-	public void deleteByUsernameAndPostId(String username, Long postId);
+	public void deleteByUsernameAndPostId(String username, int postId);
 
 }
